[feat] functors and displayed functors
I chose to define directly the category of functors, and to _let `⇒`
stand for the category of functors_. The rationale is that the
combinatorics of categorical notation should resemble that of usual
set-theoretic (type-theoretic) notation.

I also introduced _indexed equality_ (displayed equality?), noted `=*`.
It would be useful to create indexed variants of `ext` and `rw`, but
that's a story for another time.