[feat] opfibrations
The source of `Opfibration` clearly is dual to `Fibration`.
I moved some lemmas to `DisplayedCategory`.
I simplified the proof of `fmap_comp` in the construction of the
pushforward and pullback functors, but I'm still not very happy about
the looks of the proofs: although they could flow pretty naturally,
their principle is hidden behind some technicalities. Resolving this
would require some automation, and it's a non-goal at the moment. So I
guess I'll have to be happy with the current looks of the proofs.